refactor: enhance replay functionality and analysis configuration
- Introduced a new replay policy system to manage board replay behavior based on the presence of move trackers.
- Updated `GameReplayer` and `GameProcessor` to utilize the new replay policy, improving clarity and flexibility in game analysis.
- Refactored analysis configuration types to separate runtime state from public analysis types, enhancing maintainability.
- Updated documentation in `AGENTS.md` and `IDEAS.md` to reflect changes in replay logic and configuration structure.
