+++ src/types/analysis.ts
@@ -1,4 +1,4 @@
-import type { Tracker, TrackerConfig } from '#types/tracker';
+import type { Tracker } from '#types/tracker';
 
 /** User-facing analysis configuration passed to {@link Chessalyzer.analyzePGN}. */
 export interface AnalysisConfig {
@@ -21,33 +21,12 @@ export interface MultithreadConfig {
     minLines?: number;
     /**
      * @deprecated Use `targetBytes` instead. Ignored by the worker-side parse path.
+     * Still used by the legacy MT path (filter / `cntGames`).
      */
     batchSize?: number;
 }
 
-/** Normalized per-config processor state (filter, game limit). */
-export interface GameProcessorConfig {
-    hasFilter: boolean;
-    filter: (game: object) => boolean;
-    cntGames: number;
-}
-
-/** Runtime tracker buckets while processing one analysis config. */
-export interface GameProcessorAnalysisConfig {
-    trackers: { move: Tracker[]; game: Tracker[] };
-    processedMoves: number;
-    processedGames: number;
-}
-
-/** Main-thread processor config including serializable tracker metadata for workers. */
-export interface GameProcessorAnalysisConfigFull extends GameProcessorAnalysisConfig {
-    config: GameProcessorConfig;
-    trackerData: { name: string; cfg: TrackerConfig; path: string }[];
-    cntReadGames: number;
-    isDone: boolean;
-}
-
-/** Counters emitted by move trackers for one game or merged batch. */
+/** Counters emitted for one game or merged batch. */
 export interface GameAndMoveCount {
     cntGames: number;
     cntMoves: number;
